REFACTOR : Section안에 들어가는 ItemDiv AlbumBox 컴포넌트로 분할
