+++ src/components/molecules/Section.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import PlayCircleIcon from '@mui/icons-material/PlayCircle';
 import { Link } from 'react-router-dom';
+import AlbumBox from '../atoms/AlbumBox';
 
 interface Data {
   title: string;
@@ -19,6 +19,7 @@ interface Props {
   data: Data;
   dataNum: number;
   show: boolean;
+  color: string;
 }
 const WrapperSection = styled.section`
   display: flex;
@@ -44,8 +45,8 @@ const TitleLink = styled(Link)`
 `;
 const TitleP = styled.p`
   font-size: 24px;
-  color: #b3b3b3;
   font-weight: 700;
+  color: ${(props) => props.color || '#b3b3b3'};
 `;
 const ShowAllLink = styled(Link)`
   color: #b3b3b3;
@@ -80,112 +81,8 @@ const GridDiv = styled.div`
     grid-gap: 12px;
   }
 `;
-const ItemDiv = styled.div`
-  background-color: #181818;
-  padding: 16px;
-  position: relative;
-  cursor: pointer;
-  border-radius: 6px;
-  transition: background-color 0.3s ease;
-  &:hover {
-    background-color: #282828;
-    > div > div {
-      opacity: 1;
-      bottom: 8px;
-      transition: all 0.3s ease;
-    }
-  }
-`;
-const ImgDiv = styled.div`
-  margin-bottom: 16px;
-  position: relative;
-`;
-const ButtonDiv = styled.div`
-  width: 48px;
-  height: 48px;
-  position: absolute;
-  z-index: 3;
-  right: 8px;
-  bottom: 0px;
-  opacity: 0;
-  transition: all 0.3s ease;
-  &:hover {
-    > button {
-      transform: scale(1.1);
-      transition: all 0.3s ease;
-    }
-  }
-`;
 
-const PlayButton = styled.button`
-  background: transparent;
-  width: 100%;
-  height: 100%;
-  border: none;
-  position: relative;
-  padding: 0px;
-  cursor: pointer;
-`;
-const BlackDiv = styled.div`
-  width: 0;
-  height: 0;
-  border-bottom: 9.6px solid transparent;
-  border-top: 9.6px solid transparent;
-  border-left: 15.3px solid black;
-  border-right: 15.3px solid transparent;
-  position: absolute;
-  top: 50%;
-  left: 69.6%;
-  transform: translate(-50%, -50%);
-`;
-const Img = styled.img`
-  width: 100%;
-  height: 100%;
-  border-radius: 4px;
-`;
-const TextDiv = styled.div`
-  display: flex;
-  flex-direction: column;
-  min-height: 62px;
-`;
-const NameP = styled.p`
-  font-size: 16px;
-  color: white;
-  font-weight: 700;
-  overflow: hidden; // 넘쳐난 텍스트 숨기기
-  text-overflow: ellipsis; //... 말줄임 효과.
-  white-space: nowrap; // 줄바꿈 없애기.
-  margin-block-start: 4px;
-  margin-block-end: 4px;
-  padding-block-end: 4px;
-`;
-const SubNameDiv = styled.div`
-  display: flex;
-  align-items: baseline;
-  color: #b3b3b3;
-  font-size: 14px;
-  line-height: 1.2;
-`;
-
-const SubNameLink = styled(Link)`
-  line-height: 1.2;
-  height: 35px;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  font-size: 14px;
-  color: #b3b3b3;
-  overflow: hidden; // 넘쳐난 텍스트 숨기기
-  text-overflow: ellipsis; //... 말줄임 효과.
-  word-break: break-all; // 영어는 필요.
-  padding-block-start: 4px;
-  text-decoration: none;
-  &:hover {
-    text-decoration: underline;
-  }
-`;
-
-const Section = ({ data, dataNum, show }: Props) => {
+const Section = ({ data, dataNum, show, color }: Props) => {
   return (
     <WrapperSection>
       <TitleDiv>
@@ -193,7 +90,7 @@ const Section = ({ data, dataNum, show }: Props) => {
           {show ? (
             <TitleLink to="section">{data.title}</TitleLink>
           ) : (
-            <TitleP>{data.title}</TitleP>
+            <TitleP color={color}>{data.title}</TitleP>
           )}
         </HDiv>
         {show && <ShowAllLink to="section">모두 표시</ShowAllLink>}
@@ -204,35 +101,7 @@ const Section = ({ data, dataNum, show }: Props) => {
             if (idx < dataNum) return item;
           })
           .map((item: Item, idx: number) => (
-            <ItemDiv key={idx}>
-              <ImgDiv>
-                <Img src={item.imgSrc} />
-                <ButtonDiv>
-                  <PlayButton>
-                    <PlayCircleIcon
-                      sx={{
-                        color: '#12d760',
-                        fontSize: '48px',
-                      }}
-                    />
-                    <BlackDiv />
-                  </PlayButton>
-                </ButtonDiv>
-              </ImgDiv>
-              <TextDiv>
-                <NameP>{item.title}</NameP>
-                <SubNameDiv>
-                  {item.artist.map((value, index) => (
-                    <>
-                      {index > 0 ? <pre>, </pre> : null}
-                      <SubNameLink to="artist" key={index}>
-                        {value}
-                      </SubNameLink>
-                    </>
-                  ))}
-                </SubNameDiv>
-              </TextDiv>
-            </ItemDiv>
+            <AlbumBox item={item} key={idx} />
           ))}
       </GridDiv>
     </WrapperSection>
